Revert adding "technology choices" to the list of protected topics
This reverts commit 083a8aff5d64b0318607e4762c0b885452e95c5d because in a
technical conference, it is more important to be able to talk freely about
the pros and cons of technology choices than it is to be welcoming to everyone.

This does not mean that it is ok to be disrespectful to people who make any
given technology choice. It just means that, unlike any of the other topics
listed as subjects for harassment, discussion of technology choices is
on-topic for a tech conference. Discussing the health risks of certain sexual
practices, for example, could easily be seen as "offensive verbal comments",
and should be avoided or, if somehow necessary, done with extreme care. I think
we do not want to impose similar limitations on discussions of, say, security
implications of certain technology choices.