Updated Hebrew translation
While at it, added a special mention of Hebrew language gender issues.

Hebrew has no separate neutral gender and uses male as the default;
the difference can be seen in every noun, verb and adjective. Writing
a text like this in a gender-neutral way would be highly awkward and
cumbersome, and writing it in the female gender can seem less than
completely professional. Therefore, we add a paragraph shortly
explaining that we deliberately chose to use the male gender, and
apologizing to females; this is a popular way of dealing with the
issue among modern Hebrew speakers.

Of course, since the issue does not exist in English, this addition was
made only in the Hebrew translation.